requirements.txt
flask
flask-socketio
anthropic
pillow
gunicorn
eventlet
